Read on roll over does not work without an additional atomic
On front() you get a pointer but do not know how long is the element in
the buffer, and whether it rolled over to the beginning or no. Yet,
that's exactly the point when you need to know whether to move the read
pointer.

It seems like you always need an atomic check in the front(), no cache.
I.e. to check some wrapCapacity_ size that points where the buffer ends.
Because there is no way to distinguish when you do not need to know the
dynamic size and when you need to check it.

Something can be done, if the max item size is limited.

	modified:   include/rigtorp/SPSCQueue.h
	modified:   src/SPSCQueueBenchmarkCoords.cpp