An option to memcpy data into the rawData queue
You can either fill the rawData queue item, or fill an array in memory
and memcpy from it into the queue. Also, if #define MEMCPY > 0, it will
memcpy from a buffer with 1 packet in memory, not the whole buffer for
n_packets.

In both cases, memcpy option turns out about twice slower than the
direct fill, i.e about 2 ops/us. Which is about the rate in YARR
benchmark with the rudimentary ringbuffers/mempools?

The perf stats of instructions and cache misses show that memcpy does
1.45 instructions more, and ins/cycle is 1.3 times worse:
1.45*1.3 = 1.9.

But, a toy-test of memcpy 24-byte packets without -O runs at about
6GB/s and 280Mops/s (op=memcpy). So, not clear what happens here?

	modified:   include/test_parsing.h
	modified:   src/SPSCQueueBenchmarkCoords.cpp